feat: fix ghost surveillance camera rotations
Fixes ghost camera rotations for non rotating cameras. On stock, static cameras will still rotate when the player is not looking at them and over a certain distance. Might also not be a bad idea to have as an option if the stock behavior is preferred.

Video showing the issue by BluntsNBeatz
https://www.youtube.com/watch?v=-tMA9nOh5Zg
https://www.youtube.com/watch?v=faCJc3CPG5Q